accelerator_project/requirements.txt
asgiref==3.7.2
Django==3.2.24
django-environ==0.11.2
django-sslserver==0.22
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
